made TestAreas DefinedRegions again

+++ src/main/java/de/zeanon/testutils/plugin/utils/backup/Backup.java
@@ -15,14 +15,16 @@ import com.sk89q.worldedit.function.operation.Operations;
 import com.sk89q.worldedit.math.BlockVector3;
 import com.sk89q.worldedit.regions.CuboidRegion;
 import com.sk89q.worldedit.session.ClipboardHolder;
+import de.zeanon.storagemanagercore.internal.base.exceptions.RuntimeIOException;
 import de.zeanon.storagemanagercore.internal.utility.basic.BaseFileUtils;
 import de.zeanon.testutils.TestUtils;
 import de.zeanon.testutils.plugin.commands.backup.BackupCommand;
+import de.zeanon.testutils.plugin.commands.testutils.TestUtilsCommand;
 import de.zeanon.testutils.plugin.utils.ConfigUtils;
 import de.zeanon.testutils.plugin.utils.InternalFileUtils;
 import de.zeanon.testutils.plugin.utils.enums.BackupMode;
+import de.zeanon.testutils.regionsystem.region.DefinedRegion;
 import de.zeanon.testutils.regionsystem.region.RegionManager;
-import de.zeanon.testutils.regionsystem.region.TestArea;
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
@@ -51,15 +53,15 @@ public abstract class Backup extends BukkitRunnable {
 			try {
 				final @NotNull String name = LocalDateTime.now().format(Backup.getFormatter());
 				@NotNull org.bukkit.World tempWorld;
-				for (final @NotNull File regionFolder : BaseFileUtils.listFolders(TestUtils.getPluginFolder().resolve("TestAreas").toRealPath().toFile())) {
+				for (final @NotNull File regionFolder : BaseFileUtils.listFolders(TestUtilsCommand.TESTAREA_FOLDER.toRealPath().toFile())) {
 					final @NotNull File backupFolder = BackupCommand.BACKUP_FOLDER.resolve(regionFolder.getName()).toFile();
 
-					final @Nullable TestArea southRegion = RegionManager.getRegion(regionFolder.getName() + "_south");
-					final @Nullable TestArea northRegion = RegionManager.getRegion(regionFolder.getName() + "_north");
+					final @Nullable DefinedRegion southRegion = RegionManager.getRegion(regionFolder.getName() + "_south");
+					final @Nullable DefinedRegion northRegion = RegionManager.getRegion(regionFolder.getName() + "_north");
 					if (southRegion != null && northRegion != null) {
 						if (this.doBackup(southRegion, northRegion)) {
 							tempWorld = southRegion.getWorld();
-							final @NotNull File folder = BackupCommand.BACKUP_FOLDER.resolve(regionFolder.getName() + "/" + this.sequence.getPath(null) + "/" + name).toFile();
+							final @NotNull File folder = BackupCommand.BACKUP_FOLDER.resolve(regionFolder.getName()).resolve(this.sequence.getPath(null)).resolve(name).toFile();
 							this.backupSide(tempWorld, southRegion, folder);
 							southRegion.setHasChanged(false);
 							this.backupSide(tempWorld, northRegion, folder);
@@ -68,23 +70,34 @@ public abstract class Backup extends BukkitRunnable {
 							Backup.this.cleanup(backupFolder);
 						}
 					} else {
-						FileUtils.deleteDirectory(regionFolder);
-						InternalFileUtils.deleteEmptyParent(regionFolder, TestUtils.getPluginFolder().resolve("TestAreas").toFile());
+						new BukkitRunnable() {
+							@Override
+							public void run() {
+								try {
+									if (regionFolder.isDirectory()) {
+										FileUtils.deleteDirectory(regionFolder);
+										InternalFileUtils.deleteEmptyParent(regionFolder, TestUtilsCommand.TESTAREA_FOLDER.toFile());
+									}
 
-						if (backupFolder.exists()) {
-							FileUtils.deleteDirectory(backupFolder);
-							InternalFileUtils.deleteEmptyParent(backupFolder, BackupCommand.BACKUP_FOLDER.toFile());
-						}
+									if (backupFolder.exists() && backupFolder.isDirectory()) {
+										FileUtils.deleteDirectory(backupFolder);
+										InternalFileUtils.deleteEmptyParent(backupFolder, BackupCommand.BACKUP_FOLDER.toFile());
+									}
+								} catch (IOException e) {
+									throw new RuntimeIOException(e);
+								}
+							}
+						}.runTaskAsynchronously(TestUtils.getInstance());
 					}
 				}
 				this.systemOutDone();
-			} catch (IOException e) {
+			} catch (IOException | RuntimeIOException e) {
 				e.printStackTrace();
 			}
 		}
 	}
 
-	public void backupSide(final @NotNull World tempWorld, final @NotNull TestArea tempRegion, final @NotNull File folder) {
+	public void backupSide(final @NotNull World tempWorld, final @NotNull DefinedRegion tempRegion, final @NotNull File folder) {
 		final @NotNull BukkitWorld bukkitWorld = new BukkitWorld(tempWorld);
 		final @NotNull CuboidRegion region = new CuboidRegion(bukkitWorld, tempRegion.getMinimumPoint().toBlockVector3(), tempRegion.getMaximumPoint().toBlockVector3());
 		final @NotNull BlockArrayClipboard clipboard = new BlockArrayClipboard(region);
@@ -113,7 +126,7 @@ public abstract class Backup extends BukkitRunnable {
 		}
 	}
 
-	public void pasteSide(final @NotNull TestArea tempRegion, final @NotNull EditSession editSession, final @NotNull File file) throws IOException, WorldEditException { //NOSONAR
+	public void pasteSide(final @NotNull DefinedRegion tempRegion, final @NotNull EditSession editSession, final @NotNull File file) throws IOException, WorldEditException { //NOSONAR
 		try (final @NotNull ClipboardReader reader = BuiltInClipboardFormat.SPONGE_SCHEMATIC.getReader(BaseFileUtils.createNewInputStreamFromFile(file))) {
 			final @NotNull Clipboard clipboard = reader.read();
 
@@ -137,5 +150,5 @@ public abstract class Backup extends BukkitRunnable {
 
 	protected abstract void systemOutDone();
 
-	protected abstract boolean doBackup(final @NotNull TestArea southRegion, final @NotNull TestArea northRegion);
+	protected abstract boolean doBackup(final @NotNull DefinedRegion southRegion, final @NotNull DefinedRegion northRegion);
 }
